requirements.txt
captum==0.7.0
h5py==3.7.0
imageio==2.19.3
javabridge==1.0.19
matplotlib==3.5.2
numpy==1.21.6
numpy==1.24.2
opencv_python==4.7.0.68
openslide_python==1.2.0
pandas==1.4.4
Pillow==9.3.0
Pillow==10.3.0
python_bioformats==4.0.7
python_javabridge==4.0.3
scikit_learn==1.0.2
scikit_learn==1.2.2
seaborn==0.11.2
Shapely==2.0.4
statsmodels==0.13.2
torch==1.9.0
torch==2.0.0
torch==2.0.1
torchstain==1.3.0
torchvision==0.13.0
torchvision==0.15.0
tqdm==4.64.1
transformers==4.30.1